requirements.txt
openai
python-dotenv
langchain
python-telegram-bot